updated fucntions for Akky and verisign
