+++ cypress/integration/landingpages/Normal Pagevamp/Italy/pricing.spec.js
