+++ cypress/integration/landingpages/Normal Pagevamp/Mexico/pricing.spec.js
