+++ cypress/integration/landingpages/Normal Pagevamp/Pagevamp/pricing.spec.js
