+++ cypress/integration/landingpages/Normal Pagevamp/Spain/pricing.spec.js
