+++ cypress/integration/landingpages/Normal Pagevamp/helper.js
