+++ cypress/integration/landingpages/campaign/Verisign/verisignHelper.js
@@ -103,7 +103,5 @@ class verisignHelper {
         let sizes = ['iphone-6','ipad-2',[1500,1000]];
         return sizes;
     }
-   
 }
-
 export default verisignHelper
